refactor(web): extract backend-config modal from App.tsx (issue #642, Phase 3 seam 1)
Verbatim code-move: KnobMenuNumber/NumberField go to
components/backend/fields.tsx (shared with DesignSession's call sites in
App.tsx); BackendConfigModal/BackendConfigProps/BSplineFields go to
components/backend/BackendConfigModal.tsx. App.tsx drops from 8736 to
8313 lines; no logic changes, only import additions/removals.
